Refactored Sidebar and IA
