+++ src/Components.ts
@@ -1,16 +1,16 @@
 // Pages
 export {default as Map} from './Pages/Map';
 export {default as Page} from './Pages/Page';
-export {default as Valorant} from './Pages/Valorant';
-export {default as CS2} from './Pages/CS2';
-export {default as Login} from './Pages/Login';
-export {default as Register} from './Pages/Register';
+export {default as Valorant} from './Pages/game/Valorant';
+export {default as CS2} from './Pages/game/CS2';
+export {default as Login} from './Pages/auth/Login';
+export {default as Register} from './Pages/auth/Register';
 
 // Components
 export {default as Header} from './Components/Header';
 export {default as Game} from './Components/Game';
-export {default as SideNav} from './Components/SideNav';
-export {default as SideNavItems } from './Components/SideNavItems';
+export {default as SideNav} from './Components/sidebar/SideNav';
+export {default as SideNavItems } from './Components/sidebar/SideNavItems';
 export {default as Footer} from './Components/Footer';
-export {default as SideNavWrapper} from './Components/SideNavWrapper';
+export {default as SideNavWrapper} from './Components/sidebar/SideNavWrapper';
 export {default as Searchbar} from './Components/Searchbar';
