--- src/Components/SideNavWrapper.tsx
@@ -1,12 +1,0 @@
-import { SideNav, SideNavItems } from "../Components";
-
-const SideNavWrapper: React.FC = () => {
-   
-   return (
-      <>
-      
-      </>
-   );
-};
-
-export default SideNavWrapper;
