+++ src/Pages/Page.tsx
@@ -1,7 +1,5 @@
 // Main page of the app
-import { Header, Game, SideNav, SideNavItems, Footer } from "../Components";
-
-import { MdOutlineSettings,  MdOutlineGamepad, MdHome } from "react-icons/md";
+import { Header, Game, Footer, SideNavWrapper } from "../Components";
 
 import cs2 from '../assets/csgo2.webp';
 import valorant from '../assets/valorant.jpg';
@@ -10,18 +8,16 @@ const Page: React.FC = () => {
    return (
       <>
          <Header />
-         <SideNav>
-            <SideNavItems icon={<MdHome size={25} />} text="Home" active />
-            <SideNavItems icon={<MdOutlineGamepad size={25} />} text=" CS 2" />
-            <SideNavItems icon={<MdOutlineGamepad size={25} />} text="Valorant" />
-            <SideNavItems icon={<MdOutlineSettings size={25} />} text="Settings" alert />
-         </SideNav>
+
+         <SideNavWrapper />
+         
          <div className="h-screen flex">
             <div className="main-content flex-col md:flex-row flex-1 flex justify-center items-center space-x-4">
                 <Game game={"CS2"} name={cs2}/>
                 <Game game={"Valorant"} name={valorant}/>
             </div>
          </div>
+         
          <Footer />
       </>
    );
