+++ src/Pages/auth/Register.tsx
@@ -1,9 +1,7 @@
 import React, { useState, FormEvent } from 'react';
 import axios from 'axios'; // Ensure Axios is installed
 import { Link } from 'react-router-dom';
-import { Header, SideNav, SideNavItems } from '../Components';
-
-import { MdOutlineSettings,  MdOutlineGamepad, MdHome } from "react-icons/md";
+import { Header, SideNavWrapper } from '../../Components';
 
 const Register: React.FC = () => {
    const [email, setEmail] = useState<string>('');
@@ -34,12 +32,9 @@ const Register: React.FC = () => {
    return (
    <>
       <Header />
-      <SideNav>
-         <SideNavItems icon={<MdHome size={25} />} text="Home" active />
-         <SideNavItems icon={<MdOutlineGamepad size={25} />} text=" CS 2" />
-         <SideNavItems icon={<MdOutlineGamepad size={25} />} text="Valorant" />
-         <SideNavItems icon={<MdOutlineSettings size={25} />} text="Settings" alert />
-      </SideNav>
+
+      <SideNavWrapper />
+      
       <div className="h-screen md:h-full md:w-1/2 lg:w-1/3 container flex flex-col mx-auto bg-white rounded-lg md:pt-12 md:my-5">
          <div className="flex justify-center w-full h-full my-auto xl:gap-14 lg:justify-normal md:gap-5 draggable">
             <div className="flex items-center justify-center w-full lg:p-12">
