+++ src/Pages/game/Valorant.tsx
@@ -1,6 +1,6 @@
-import { Header, SideNav, SideNavItems, Searchbar, Footer } from "../Components";
+import { Header, SideNav, SideNavItems, Searchbar, Footer } from "../../Components";
 import { MdOutlineSettings, MdOutlineGamepad, MdHome } from "react-icons/md";
-import valorantbanner from '../assets/valorantbanner.webp';
+import valorantbanner from '../../assets/valorantbanner.webp';
 
 const Valorant: React.FC = () => {
    const handleSearch = (searchTerm: string) => {
